refactor: Move middleware to separate file
